refactor(MEMBER): 회원 정보 생성, 로그인 기능 AcceptanceDataGenerator class로 이관
